useless calls of mir_*strlen in DrawText replaced with -1
git-svn-id: http://svn.miranda-ng.org/main/trunk@11223 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/GTalkExt/src/handlers.cpp
@@ -45,7 +45,7 @@ GoogleTalkAcc* isGoogle(LPARAM lParam)
 void FormatMessageUrl(LPCTSTR format, LPTSTR buf, LPCTSTR mailbox, LPCTSTR tid)
 {
 	ULARGE_INTEGER iTid; iTid.QuadPart = _tstoi64(tid);
-	int l = mir_tstrlen(buf);
+	size_t l = mir_tstrlen(buf);
 	mir_sntprintf(buf, l, format, mailbox, iTid.HighPart, iTid.LowPart);
 	assert(l >= mir_tstrlen(buf));
 }
